+++ src/model/logic/Geometry.java
@@ -6,9 +6,9 @@ public class Geometry {
 
 	public String type;
 	
-	public ArrayList<Integer> coordinates;
+	public ArrayList<Double> coordinates;
 	
-	public Geometry(String type, ArrayList<Integer> coordinates){
+	public Geometry(String type, ArrayList<Double> coordinates){
 		
 		this.type = type;
 		this.coordinates = coordinates;
